table format

+++ src/app/siswa/page.jsx
@@ -2,14 +2,16 @@
 
 import { useEffect, useState } from "react"
 import Index from "../index"
-import Table from "@/components/Table"
 import UploadSheet from "@/components/UploadSheet"
 import InputData from "@/components/InputData"
 import Create from "./create"
 import axios from "axios"
-import { getUrl, kelasFormat, siswaFormat } from "../../../utils/format"
+import { getUrl, kelasFormat, getNum, getKelas, deleteDialog } from "../../../utils/format"
 import Pagination from "@/components/Pagination"
-import { useSearchParams } from "next/navigation"
+import TableFormat, { Td, Tr, Link, Button } from "@/components/TableFormat"
+import { useSelector } from "react-redux"
+import { getId } from "@/redux/features/tahunAjarSlice"
+import { Delete, Edit } from "../../../public/svg"
 
 export default function Siswa() {
   const [dataSiswa, setDataSiswa] = useState([])
@@ -19,18 +21,18 @@ export default function Siswa() {
   const [index, setIndex] = useState(0)
   const [page, setPage] = useState(1)
   const [total, setTotal] = useState(0)
-  const searchParams = useSearchParams()
   const url = "/api/siswa/"
+
+  const tahunId = useSelector(getId)
 
   const fetchData = async () => {
     try {
       const res = await axios.get(getUrl(`/api/siswa?page=${page}`))
       const resKelas = await axios.get(getUrl('/api/kelas?page=all'))
       const formattedKelas = await kelasFormat(resKelas.data.kelas)
+      setDataSiswa(res.data.siswa)
       setKelas(formattedKelas)
       setTotal(res.data.total)
-      const formattedData = await siswaFormat(res.data.siswa, page, searchParams.get('tahun'))
-      setDataSiswa(formattedData)
     } catch (err) {
       console.error(err)
       setError(err.message)
@@ -63,6 +65,10 @@ export default function Siswa() {
   }
 
   const deleteHandler = async (id) => {
+    if(!deleteDialog()) {
+      return;
+    }
+
     setLoading(true)
     await axios.delete(getUrl(`${url}${id}`))
       .then(res => console.log(res))
@@ -76,7 +82,7 @@ export default function Siswa() {
   useEffect(() => {
     setLoading(true)
     fetchData()
-  }, [page, searchParams.get('tahun')])
+  }, [page])
 
   return (
     <Index title='Siswa' placeholder='Cari Siswa (NIS, Nama)...'>
@@ -87,7 +93,28 @@ export default function Siswa() {
           <Create loading={loading} submitHandler={submitHandler} kelas={kelas} />
         </InputData>
       </div>
-      <Table title='Akun Siswa' data={dataSiswa} loading={loading} error={error} viewHandler={'/siswa/'} editHandler={editHandler} deleteHandler={deleteHandler} />
+      <TableFormat title='Akun Siswa' format={['No', 'Nama Siswa', 'Kelas', 'NIS', 'NISN', 'JK', 'Alamat', 'HP', 'Data Dibuat', 'Data Diubah']} loading={loading} error={error}>
+        {
+          dataSiswa.map((siswa, index) => (
+            <Tr>
+              <Td>{getNum(page, index)}</Td>
+              <Td><Link title={siswa.nama}/></Td>
+              <Td><Link title={getKelas(siswa.kelas, tahunId)?.namaKelas}/></Td>
+              <Td><Link title={siswa.nis}/></Td>
+              <Td><Link title={siswa.nisn}/></Td>
+              <Td><Link title={siswa.jk}/></Td>
+              <Td><Link title={siswa.alamat}/></Td>
+              <Td><Link title={siswa.hp}/></Td>
+              <Td><Link title={siswa.createdAt}/></Td>
+              <Td><Link title={siswa.updatedAt}/></Td>
+              <Td className='flex flex-row flex-wrap gap-2'>
+                <Button backgroundColor={'bg-orange-500'}><Edit/></Button>
+                <Button clickHandler={() => deleteHandler(siswa.id)} backgroundColor={'bg-red-500'}><Delete/></Button>
+              </Td>
+            </Tr>
+          ))
+        }
+      </TableFormat>
       <Pagination page={page} setPage={setPage} loading={loading} total={total} />
     </Index>
   )
